added interface for services and renamed any classes
