+++ src/main/java/ru/neustupov/advpost/Application.java
@@ -6,10 +6,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @EnableAdminServer
 @SpringBootApplication
-public class AdvPostApplication {
+public class Application {
 
     public static void main(String[] args) {
-        SpringApplication.run(AdvPostApplication.class, args);
+        SpringApplication.run(Application.class, args);
     }
 
 }
