+++ src/main/java/ru/neustupov/advpost/ApplicationConfig.java
@@ -5,5 +5,5 @@ import org.springframework.context.annotation.PropertySource;
 
 @Configuration
 @PropertySource("classpath:application.yaml")
-public class AppConfig {
+public class ApplicationConfig {
 }
